[wpilib] Remove GyroBase
