+++ wpilibc/src/main/native/include/frc/romi/RomiGyro.h
@@ -6,7 +6,10 @@
 
 #include <hal/SimDevice.h>
 
-#include "frc/GyroBase.h"
+#include "frc/ErrorBase.h"
+#include "frc/interfaces/Gyro.h"
+#include "frc/smartdashboard/Sendable.h"
+#include "frc/smartdashboard/SendableHelper.h"
 
 namespace frc {
 
@@ -16,7 +19,10 @@ namespace frc {
  * This class is for the Romi onboard gyro, and will only work in
  * simulation/Romi mode. Only one instance of a RomiGyro is supported.
  */
-class RomiGyro : public GyroBase {
+class RomiGyro : public Gyro,
+                 public ErrorBase,
+                 public Sendable,
+                 public SendableHelper<RomiGyro> {
  public:
   RomiGyro();
 
@@ -83,6 +89,8 @@ class RomiGyro : public GyroBase {
    * Resets the gyro
    */
   void Reset() override;
+
+  void InitSendable(SendableBuilder& builder) override;
 
  private:
   hal::SimDevice m_simDevice;
